\name{numeps}
\alias{numeps}
\docType{data}
\title{
internal
}
\description{
internal use only}
\usage{data(numeps)}
\format{
  The format is:
 num 1e-25
}
\details{
internal use only
}
\source{
internal use only
}
\references{
internal use only
}
\examples{
data(numeps)
## maybe str(numeps) ; plot(numeps) ...
}
\keyword{internal}
