\name{mgpd-package}
\alias{mgpd-package}
\alias{mgpd}
\docType{package}
\title{
Functions for multivariate generalized Pareto distribution (MGPD of Type II) 
}
\description{
Extends density functions to parametric multivariate generalized Pareto distributions (MGPD of Type II), 
and provides fitting functions which calculate maximum likelihood estimates for bivariate models (BGPD). 
There are seven "classical" parametric dependence models within BGPD available: logistic, negative logistic, bilogistic, 
negative bilogistic, Tajvidi's generalised symmetric logistic, Dirichlet and mixed. Beyond these, two novel 
asymmetric approaches giving absolutely continuous BGPD are implemented as \emph{psi-logistic} and \emph{psi-negative logistic} models.
}
\details{
\tabular{ll}{
Package: \tab mgpd\cr
Type: \tab Package\cr
Version: \tab 1.0 (bgpd version)\cr
Date: \tab 2011-03-09 \cr
}
}

\author{
Pal Rakonczai <rakonczai.p@gmail.com>
}
\seealso{
\code{\link{dbgpd}}, \code{\link{dbgpd_region}}, \code{\link{bgpd_fit}}, \code{\link{DailyWind}}
}

